requirements.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.20.0
opencv-python>=4.5.0
einops>=0.4.1
Pillow>=8.0.0
numpy>=1.19.0 